Changed API of random module to use size_t
size_t is used as the type of the amount of generated bytes.

Also, fixed a typo on include guard.

+++ src/onion/random-default.c
@@ -31,7 +31,7 @@
 #include "types_internal.h"
 #include "log.h"
 
-static unsigned onion_random_refcount=0;
+static size_t onion_random_refcount=0;
 
 /**
  * @short Initializes the global random number generator
@@ -80,9 +80,9 @@ void onion_random_free() {
  *
  * Generate size bytes of random data and put on data
  */
-void onion_random_generate(void* data, unsigned int size) {
+void onion_random_generate(void* data, size_t size) {
 	unsigned char* data_char = data;
-	unsigned int i;
+	size_t i;
 	for( i=0;i<size;++i ) {
 		data_char[i] = rand();
 	}
